[Change] EditProfileBtn split component
